+++ src/main.rs
@@ -1,21 +1,15 @@
 mod error;
 mod message;
+mod router;
 
+use crate::router::api::api;
+
+use axum::Router;
+use serde::Deserialize;
 use std::sync::Arc;
 
-use error::ServerError;
-use message::LineSendMessege;
-
-use axum::{
-    response::IntoResponse,
-    routing::{get, post},
-    Extension, Router,
-};
-use reqwest::StatusCode;
-use serde::Deserialize;
-
 #[derive(Debug)]
-struct State {
+pub struct State {
     token: String,
 }
 
@@ -31,31 +25,11 @@ async fn main() -> Result<(), error::ServerError> {
     let state = Arc::new(State {
         token: config.access_token,
     });
-
-    let app = Router::new()
-        .route(
-            "/",
-            get(|| async {
-                println!("get /");
-                "Hello, World!"
-            }),
-        )
-        .route("/send", post(send))
-        .layer(Extension(state));
-
+    let app = Router::new().nest("/api/v1", api(state));
     let listener = tokio::net::TcpListener::bind("0.0.0.0:8001").await?;
+
     println!("* start server 0.0.0.0:8001");
     axum::serve(listener, app).await?;
 
     Ok(())
-}
-
-async fn send(
-    Extension(state): Extension<Arc<State>>,
-    body: String,
-) -> Result<impl IntoResponse, ServerError> {
-    let message = LineSendMessege::new(body);
-    message.send_message(state.token.clone()).await?;
-    println!("send message");
-    Ok(StatusCode::OK)
 }
