+++ src/router.rs
@@ -1,0 +1,1 @@
+pub mod api;
